added 2 commands send mail, and some refactored    - added Send(Anonymously/CurrentMail)Command    - refactored prepare CallbackData for some commands    - change and added methods ModifyDataBaseServiceImpl    - refactored UnknownCommand    - added command in CommandContainer    - added methods for Utils

+++ src/main/java/org/gnori/mailsenderbot/dto/MessageSentRecordDto.java
@@ -6,7 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import java.time.LocalDateTime;
 
 public class MessageSentRecordDto {
-    private final Long countMessages;
+    private final Integer countMessages;
     private final String sendDate;
     public MessageSentRecordDto(MessageSentRecord messageSentRecord) {
         this.countMessages = messageSentRecord.getCountMessages();
@@ -14,7 +14,7 @@ public class MessageSentRecordDto {
 
     }
 
-    public Long getCountMessages() {
+    public Integer getCountMessages() {
         return countMessages;
     }
